afxdp/parser: relocate inline tests to parser_tests.rs
Same playbook. 162-LOC inline test block extracted. parser.rs: 339 → 180 LOC.
cargo build/test 92/878.
